+++ wemake_python_styleguide/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Contains all files related to WPS CLI utility."""