+++ wemake_python_styleguide/cli/commands/explain/__init__.py
@@ -1,0 +1,1 @@
+"""Contains files related to wps explain command."""