+++ wemake_python_styleguide/cli/commands/explain/message_formatter.py
@@ -1,6 +1,10 @@
+"""Provides tools for formatting explanations."""
+
 from typing import Final
 
-from wemake_python_styleguide.cli.commands.explain.violation_loader import ViolationInfo
+from wemake_python_styleguide.cli.commands.explain.violation_loader import (
+    ViolationInfo,
+)
 
 _DOCS_URL: Final = (
     'https://wemake-python-styleguide.readthedocs.io/en/latest/pages/'
@@ -9,14 +13,25 @@ _DOCS_URL: Final = (
 
 
 def _clean_text(text: str) -> str:
-    return text.replace("\r\n", "\n").replace("\r", "\n")
+    """
+    Cleans provided text.
+
+    Args:
+        text: target text
+
+    Returns:
+        text with normalized newlines (CRs and CRLFs transformed to LFs).
+    """
+    return text.replace('\r\n', '\n').replace('\r', '\n')
 
 
 def _replace_tabs(text: str, tab_size: int = 4) -> str:
-    return text.replace("\t", " " * tab_size)
+    """Replace all tabs with defined amount of spaces."""
+    return text.replace('\t', ' ' * tab_size)
 
 
 def _get_whitespace_prefix(line: str) -> int:
+    """Get length of whitespace prefix of string."""
     for char_index, char in enumerate(line):
         if char != ' ':
             return char_index
@@ -24,10 +39,11 @@ def _get_whitespace_prefix(line: str) -> int:
 
 
 def _get_greatest_common_indent(text: str) -> int:
-    lines = text.split("\n")
+    """Get the greatest common whitespace prefix length of all lines."""
+    lines = text.split('\n')
     if len(lines) == 0:
         return 0
-    greatest_common_indent = float("+inf")
+    greatest_common_indent = float('+inf')
     for line in lines:
         if len(line.strip()) == 0:
             continue
@@ -35,25 +51,27 @@ def _get_greatest_common_indent(text: str) -> int:
             greatest_common_indent,
             _get_whitespace_prefix(line)
         )
-    if greatest_common_indent == float("+inf"):
+    if isinstance(greatest_common_indent, float):
         greatest_common_indent = 0
     return greatest_common_indent
 
 
 def _remove_indentation(text: str, tab_size: int = 4) -> str:
+    """Remove excessive indentation."""
     text = _replace_tabs(_clean_text(text), tab_size)
     max_indent = _get_greatest_common_indent(text)
-    return "\n".join(line[max_indent:] for line in text.split("\n"))
+    return '\n'.join(line[max_indent:] for line in text.split('\n'))
 
 
 def format_violation(violation: ViolationInfo) -> str:
+    """Format violation information."""
     cleaned_docstring = _remove_indentation(violation.docstring)
     violation_url = _DOCS_URL.format(
         violation.section,
         violation.fully_qualified_id,
     )
     return (
-        f"WPS{violation.code} ({violation.identifier})\n"
-        f"{cleaned_docstring}\n"
-        f"See at website: {violation_url}"
+        f'WPS{violation.code} ({violation.identifier})\n'
+        f'{cleaned_docstring}\n'
+        f'See at website: {violation_url}'
     )